% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SCESet-methods.R
\name{addfData}
\alias{addfData}
\title{Add columns to featureData for an SCESet object}
\usage{
addfData(x, df)
}
\arguments{
\item{x}{An SCESet object.}

\item{df}{a data.frame to add to phenoData}
}
\value{
An AnnotatedDataFrame that can be assigned to \code{pData(x)}.
}
\description{
SCESet objects contain feature (i.e. gene) information (inherited from the
ExpressionSet class). This function allows convenient addition of one or more
columns to the feature data, and returns an AnnotatedDataFrame that can be
assigned to \code{fData(x)}.
}
\examples{
\dontrun{

}
}

