% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SCESet-methods.R
\name{addpData}
\alias{addpData}
\title{Add columns to phenoData for an SCESet object}
\usage{
addpData(x, df)
}
\arguments{
\item{x}{An SCESet object.}

\item{df}{a data.frame to add to phenoData}
}
\value{
An AnnotatedDataFrame that can be assigned to \code{pData(x)}.
}
\description{
SCESet objects contain phenotype information (inherited from the
ExpressionSet class). This function allows convenient addition of one or more
columns to the phenotype data, and returns an AnnotatedDataFrame.
}
\examples{
\dontrun{

}
}

