% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aggregateResults.R
\name{aggregateResults}
\alias{aggregateResults}
\title{Aggregate and compare results from different ranking methods using exact rank-product tests}
\usage{
aggregateResults(results, nmax_genes = 500, max_rankprod_exact = 1e+05)
}
\arguments{
\item{results}{a matrix or data.frame object where columns provide
results values (e.g. p-values) to be used for ranking features
(e.g. genes) and ordering of results.}

\item{nmax_genes}{integer giving the maximum number of features
(e.g. genes) for which to compute exact rank-product p-values}

\item{max_rankprod_exact}{integer giving the maximum rank-product for
which to compute exact p-values. Exact p-values are very slow to
compute for large rank-products, and converge to approximate
p-values.}
}
\value{
data frame with mean rank, rank product, number of ways of
obtaining given rank product, approximate p-value and exact p-value
}
\description{
Aggregate and compare results from different ranking methods using exact rank-product tests
}

