% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rank_product_exact.R
\name{bincoeff}
\alias{bincoeff}
\title{Calculate binomial coefficient}
\usage{
bincoeff(n, k)
}
\arguments{
\item{n}{integer giving total number of items}

\item{k}{integer giving number of items to choose}
}
\value{
integer value for the binomial coefficient
}
\description{
Calculate the binomial coefficient.
}
\examples{
bincoeff(10, 2)
}
\seealso{
choose
}

