% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calcIsExpr.R
\name{calcIsExpr}
\alias{calcIsExpr}
\title{Calculate which gene are expressed in which cells using a count
threshold}
\usage{
calcIsExpr(object, lowerDetectionLimit = NULL, expr_data = "counts")
}
\arguments{
\item{object}{an SCESet object with expression and/or count data}

\item{lowerDetectionLimit}{numeric scalar giving the minimum expression level
for an expression observation in a cell for it to qualify as expressed}

\item{expr_data}{character scalar indicating whether the count data ("counts") or the
transformed expression data ("exprs") should be used to define if an
observation is expressed or not}
}
\value{
a logical matrix indicating whether or not a gene in a particular
cell is expressed.
}
\description{
Calculate which gene are expressed in which cells using a count
threshold
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
example_sceset <- newSCESet(countData = sc_example_counts)
isExpr(example_sceset) <- calcIsExpr(example_sceset, lowerDetectionLimit = 1,
expr_data = "exprs")
}

