% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calcIsExpr.R
\name{calcIsExprCounts}
\alias{calcIsExprCounts}
\title{Calculate which gene are expressed in which cells using a count
threshold}
\usage{
calcIsExprCounts(object, min_gene_count_in_cell = 5)
}
\arguments{
\item{object}{a DGEList object with count data}

\item{min_gene_count_in_cell}{integer giving the minimum gene count
in a cell for it to qualify as expressed}
}
\value{
a DGEList object with an 'isexpr_counts' object added, a
matrix of TRUE/FALSE values indicating whether or not a gene in a
particular cell is expressed.
}
\description{
Calculate which gene are expressed in which cells using a count
threshold
}
\examples{
d <- edgeR::DGEList(counts = matrix(rpois(100, lambda = 20), nrow = 10))
calcIsExprCounts(d)
}

