% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calcIsExprCPM}
\alias{calcIsExprCPM}
\title{Calculate which gene are expressed in which cells using a
counts-per-million threshold}
\usage{
calcIsExprCPM(object, min_gene_count_in_cell = 5,
  normalized.lib.sizes = FALSE)
}
\arguments{
\item{object}{a DGEList object with count data}

\item{min_gene_count_in_cell}{integer giving the minimum gene count
in a cell for it to qualify as expressed}

\item{normalized.lib.sizes}{logical passed to cpm(); whether or not
to use normalized library sizes to compute counts-per-million}
}
\value{
a DGEList object with an 'isexpr_cpm' object added, a
matrix of TRUE/FALSE values indicating whether or not a gene in a
particular cell is expressed.
}
\description{
Calculate which gene are expressed in which cells using a
counts-per-million threshold
}
\details{
Could incorporate normalization factors when computing
counts-per-million, but this is not implemented yet.
}
\examples{
d <- DGEList(counts = matrix(rpois(100, lambda = 20), nrow = 10))
d$cells.info <- data.frame(treatment = rep(c("A", "B"), each = 5), lib_size = colSums(d$counts))
calcIsExprCPM(d)
}

