% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/compareResults.R, R/testRankProduct.R
\name{calcRankProdPVals}
\alias{calcRankProdPVals}
\title{Calculate exact rank-product p-values}
\usage{
calcRankProdPVals(ranks, nmax.genes = 500, max.rprod.exact = 1e+05)

calcRankProdPVals(ranks, nmax.genes = 500, max.rprod.exact = 1e+05)
}
\arguments{
\item{ranks}{a matrix of logFC-ranks, one row per gene, one column
per "experiment", i.e. pair of samples}

\item{nmax.genes}{integer giving the maximum number of genes for
which to compute exact rank-product p-values}

\item{max.rprod.exact}{integer giving the maximum rank-product for
which to compute exact p-values. Exact p-values are very slow to
compute for large rank-products, and converge to approximate
p-values.}

\item{ranks}{a matrix of logFC-ranks, one row per gene, one column
per "experiment", i.e. pair of samples}

\item{nmax.genes}{integer giving the maximum number of genes for
which to compute exact rank-product p-values}

\item{max.rprod.exact}{integer giving the maximum rank-product for
which to compute exact p-values. Exact p-values are very slow to
compute for large rank-products, and converge to approximate
p-values.}
}
\value{
data frame with mean rank, rank product, number of ways of
obtaining given rank product, approximate p-value and exact p-value

data frame with mean rank, rank product, number of ways of
obtaining given rank product, approximate p-value and exact p-value
}
\description{
Calculate exact rank-product p-values

Calculate exact rank-product p-values
}
\examples{
ptm <- proc.time()
ranks <- replicate(3, sample(50))
rownames(ranks) <- paste0("gene.", 1:50)
colnames(ranks) <- paste0("exp.", 1:3)
results <- calcRankProdPVals(ranks)
results
proc.time() - ptm
ptm <- proc.time()
ranks <- replicate(3, sample(50))
rownames(ranks) <- paste0("gene.", 1:50)
colnames(ranks) <- paste0("exp.", 1:3)
results <- calcRankProdPVals(ranks)
results
proc.time() - ptm
}

