% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/compareResults.R, R/testRankProduct.R
\name{calcRankProdPvalsApprox}
\alias{calcRankProdPvalsApprox}
\title{Right-tailed gamma approximation to rank-product p-value}
\usage{
calcRankProdPvalsApprox(r, k, n)

calcRankProdPvalsApprox(r, k, n)
}
\arguments{
\item{r}{integer rank-product}

\item{k}{integer number of experiments}

\item{n}{integer number of genes}

\item{r}{integer rank-product}

\item{k}{integer number of experiments}

\item{n}{integer number of genes}
}
\value{
float aproximate p-value

float aproximate p-value
}
\description{
Right-tailed gamma approximation to rank-product p-value

Right-tailed gamma approximation to rank-product p-value
}
\examples{
calcRankProdPvalsApprox(9720, 5, 500)
calcRankProdPvalsApprox(9720, 5, 500)
}

