% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calcRankProdPVals.R
\name{calcRankProdPValsBounds}
\alias{calcRankProdPValsBounds}
\title{Calculate bounds on p-values for rank-product statistics}
\usage{
calcRankProdPValsBounds(rho, n, k, delta = "geometric")
}
\arguments{
\item{rho}{a vector of integers corresponding to the rank products
for which one wishes to compute the p-value.}

\item{n}{the number of features (e.g. genes).}

\item{k}{the number of replicates.}

\item{delta}{a character string indicating whether an upper bound
('upper'), lower bound ('lower'), or the default,geometric
approximation (i.e. the geometric mean of the upper and lower
bounds) ('geometric') should be computed.}
}
\value{
a vector of p-values, one for each rank product.
}
\description{
This function computes bounds on the p-value for rank products.
}
\details{
The exact p-value is guaranteed to be in between the lower
and the upper bound. The geometric mean of the two bounds can be
used as an approximation. Each bound is a piecewise continuous
function of the rank product. The different pieces each have an
analytic form, the parameters of which can be computed recursively.

This implementation closely follows the description in Heskes,
Eisinga, Breitling: "A fast algorithm for determining bounds and
accurate approximate p-values of the rank product statistic for
replicate experiments", further referred to as HEB. More
specifically, this R function corresponds to the recursive variant,
sketched as pseudocode in the additional material of HEB.

We thank HEB for making their R code available. Their original code
has been lightly modified for use in \pkg{scater}.
}
\examples{
calcRankProdPValsBounds(c(10, 100, 1000), 2000, 10, "upper")
}

