% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qc.R
\name{calculateQCMetrics}
\alias{calculateQCMetrics}
\title{Calculate QC metrics}
\usage{
calculateQCMetrics(object, control_genes = NULL, nmads = 5)
}
\arguments{
\item{object}{an SCESet object containing expression values and
experimental information. Must have been appropriately prepared.}

\item{control_genes}{a character vector of gene names, or a logical vector,
or a numeric vector of indices used to identify control genes (for example,
ERCC spike-in genes)}

\item{nmads}{numeric scalar giving the number of median absolute deviations to be
used to flag potentially problematic cells based on depth (total number of
counts for the cell, or library size) and coverage (number of genes with
non-zero expression). Default value is 5.}
}
\description{
Calculate QC metrics
}
\details{
Calculate useful quality control metrics to help with pre-processing
of data and identification of potentially problematic genes and cells.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
pd <- new("AnnotatedDataFrame", data=sc_example_cell_info)
example_sceset <- newSCESet(countData=sc_example_counts, phenoData=pd)
example_sceset <- calculateQCMetrics(example_sceset)
}

