% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calculate-expression.R
\name{calculateTPM}
\alias{calculateTPM}
\title{Calculate transcripts-per-million (TPM)}
\usage{
calculateTPM(object, effective_length = NULL, calc_from = "counts")
}
\arguments{
\item{object}{an \code{SCESet} object}

\item{effective_length}{vector of class \code{"numeric"} providing the
effective length for each feature in the \code{SCESet} object}

\item{calc_from}{character string indicating whether to compute TPM from
\code{"counts"}, \code{"norm_counts"}, \code{"fpkm"} or \code{"norm_fpkm"}.
Default is to use \code{"counts"}, in which case the \code{effective_length}
argument must be supplied.}
}
\value{
Matrix of TPM values.
}
\description{
Calculate transcripts-per-million (TPM) values for expression from counts for
a set of features.
}
\examples{
\dontrun{
tpm(object) <- calculateTPM(object, effective_length, calc_from="counts")
}
}

