% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qc.R
\name{findImportantPCs}
\alias{findImportantPCs}
\title{Find most important principal components for a given variable}
\usage{
findImportantPCs(object, variable = "coverage")
}
\arguments{
\item{object}{an SCESet object containing expression values and
experimental information. Must have been appropriately prepared.}

\item{variable}{character scalar providing a variable name (column from
\code{pData(object)}) for which to determine the most important PCs.}
}
\description{
Find most important principal components for a given variable
}
\details{
Plot the top 5 most important PCs for a given variable. Importance
here is defined as the average silhouette width (computed with
\code{cluster::silhouette}) when cells are clustered by the values of the
variable. If the variable is discrete, unique values of the variable define
the clusters. If the variable is continuous, it is coerced into three
discrete values, "low" (bottom quartile), "medium" (middle two quartiles)
and "high" (upper quartile), and these levels are used to define clusters.
Cells are coloured by their cluster in the plot.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
pd <- new("AnnotatedDataFrame", data = sc_example_cell_info)
example_sceset <- newSCESet(countData = sc_example_counts, phenoData = pd)
example_sceset <- calculateQCMetrics(example_sceset)
findImportantPCs(example_sceset, variable="coverage")
}

