% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SCESet-methods.R
\docType{methods}
\name{genePairwiseDistances}
\alias{geneDist}
\alias{geneDist,SCESet-method}
\alias{geneDist.SCESet}
\alias{geneDist<-}
\alias{geneDist<-,SCESet,matrix-method}
\alias{genePairwiseDistances}
\alias{genePairwiseDistances,SCESet-method}
\alias{genePairwiseDistances.SCESet}
\alias{genePairwiseDistances<-}
\alias{genePairwiseDistances<-,geneDist}
\title{genePairwiseDistances in an SCESet object}
\usage{
genePairwiseDistances(object)

genePairwiseDistances(object) <- value

geneDist(object)

geneDist(object) <- value

genePairwiseDistances.SCESet(object)

\S4method{genePairwiseDistances}{SCESet}(object)

geneDist.SCESet(object)

\S4method{geneDist}{SCESet}(object)

\S4method{genePairwiseDistances}{SCESet,matrix}(object) <- value

\S4method{geneDist}{SCESet,matrix}(object) <- value
}
\arguments{
\item{object}{a \code{SCESet} object.}

\item{value}{a matrix of class \code{"numeric"} containing gene pairwise
distances}

\item{object}{An \code{SCESet} object from which to get genePairwiseDistances.}
}
\value{
An SCESet object containing new gene pairwise distances matrix.
}
\description{
SCESet objects can contain a matrix of pairwise distances between genes. These
 functions conveniently access and replace the gene pairwise distances with the value
 supplied, which must be a matrix of the correct size. The function \code{geneDist}
 is simply shorthand for \code{genePairwiseDistances}.


}
\examples{
\dontrun{

}
}
\author{
Davis McCarthy
}

