% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getRanksOneBlock}
\alias{getRanksOneBlock}
\title{Get ranks from logFCs for one block}
\usage{
getRanksOneBlock(logFC_matrix, group)
}
\arguments{
\item{logFC_matrix}{numeric matrix of logFCs computed as group1 -
group2}

\item{group}{character vector or factor defining the experimental
groups being compared}
}
\description{
Get ranks both for up-regulation in group1 and up-regulation in
group2
}

