% Generated by roxygen2 (4.0.2): do not edit by hand
\name{makeUnique}
\alias{makeUnique}
\title{Update the parameters so that they are unique}
\usage{
makeUnique(param)
}
\arguments{
\item{param}{a single list with values for the parameters a through
e; these parameters specify the functional form of the bound; a, b,
and c are all vectors of unknown length, d and e are scalars.}
}
\value{
A possibly updated and then more concise set of parameters
containing only unique combinations of the parameters a and b.
}
\description{
This subroutine updates the parameters for a specific number of replicates and interval
such that it contains only unique combinations of the parameters a and b.
}
\details{
While updating the vectors a and b, one may end up with
the exact same combinations of a and b. Given the functional form
of the bound, the representation can then be made more concise by
simply adding the corresponding elements of c.
}

