% Generated by roxygen2 (4.0.2): do not edit by hand
\name{multcoeff}
\alias{multcoeff}
\title{Calculate multinomial coefficient}
\usage{
multcoeff(n, x)
}
\arguments{
\item{n}{integer giving total number of items}

\item{x}{vector of number of items in each of the multinomial classes}
}
\value{
integer value for the multinomial coefficient
}
\description{
Calculate multinomial coefficient
}
\examples{
multcoeff(10, c(2, 1, 3))
}
\seealso{
choose
}

