% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SCESet-methods.R
\name{newSCESet}
\alias{newSCESet}
\title{Creates a new SCESet object.}
\usage{
newSCESet(cellData = NULL, phenoData = NULL, featureData = NULL,
  experimentData = NULL, countData = NULL, lowerDetectionLimit = 0,
  logged = FALSE, isExprs = NULL)
}
\arguments{
\item{cellData}{expression data matrix for an experiment}

\item{phenoData}{data frame containing attributes of individual cells}

\item{featureData}{data frame containing attributes of features (e.g. genes)}

\item{experimentData}{MIAME class object containing metadata data and details
about the experiment and dataset.}

\item{countData}{data matrix containing raw count expression values}

\item{lowerDetectionLimit}{the minimum expression level that constitutes true
expression (defaults to zero and uses count data to determine if an
observation is expressed or not)}

\item{logged}{logical, if a value is supplied for the cellData argument, are
the expression values already on the log2 scale, or not?}

\item{isExprs}{matrix of class \code{"logical"}, indicating whether
or not each observation is above the \code{lowerDetectionLimit}.}
}
\value{
a new SCESet object
}
\description{
Scater requires that all data be housed in SCESet objects. SCESet extends
Bioconductor's ExpressionSet class, and the same basic interface is
supported. newSCESet() expects a matrix of expression values as its first
argument, with rows as features (usually genes) and columns as cells.
Per-feature and per-cell metadata can be supplied with the featureData and
phenoData arguments, respectively. Use of these optional arguments is
strongly encouraged. The SCESet also includes a slot 'counts' to store an
object containing raw count data.
}
\details{
SCESet objects store a matrix of expression values. These values are
typically counts-per-million (cpm), fragments per kilobase per million mapped
(FPKM) or some other output from a program that calculates expression values
from RNA-Seq reads. However, they could also be values from a single cell
qPCR run or some other type of assay. The newSCESet function can also accept
raw count values, in which case it uses a function from the package edgeR to
compute log2(counts-per-million) and uses these as the expression values.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
pd <- new("AnnotatedDataFrame", data = sc_example_cell_info)
example_sceset <- newSCESet(countData = sc_example_counts, phenoData = pd)
example_sceset
}

