% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{pData<-}
\alias{pData<-}
\title{Replacement for phenoData}
\usage{
pData(x) <- value
}
\arguments{
\item{x}{object containing phenoData to replace}

\item{value}{AnnotatedDataFrame to replace existing phenoData}
}
\description{
Replacement for phenoData
}

