% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SCESet-methods.R
\docType{methods}
\name{pData<-,SCESet,AnnotatedDataFrame-method}
\alias{pData<-,SCESet,AnnotatedDataFrame-method}
\title{Replaces phenoData in an SCESet object}
\usage{
\S4method{pData}{SCESet,AnnotatedDataFrame}(x) <- value
}
\arguments{
\item{x}{An SCESet object.}

\item{value}{an AnnotatedDataFrame with updated phenoData to replace
existing}
}
\value{
A matrix of expression count data, where rows correspond to features
(e.g. genes) and columns correspond to cells.
}
\description{
SCESet objects contain phenotype information (inherited from the
ExpressionSet class). This function conveniently replaces the feature data
with the value supplied, which must be an AnnotatedDataFrame.
}
\examples{
\dontrun{

}
}

