% Generated by roxygen2 (4.0.2): do not edit by hand
\name{piltzcount}
\alias{piltzcount}
\title{Number of ways to construct a rank-product}
\usage{
piltzcount(r, k, n)
}
\arguments{
\item{r}{integer rank-product value}

\item{k}{integer number of experiments}

\item{n}{integer number of genes}
}
\value{
integer value for the number of ways to construct the rank-product
}
\description{
Compute the number of k-tuples of rank-product value r, given n genes, which is equal to the number of ways a rank-product r can be constructed from k experiments if n is the number of genes (and thus the maximum rank in a single experiment)
}
\examples{
#--- Example 1.
# Calculate for n=500 the number of (k=)5-tuples with rankproduct rp=9720.
piltzcount(9720, 5, 500)

# Gamma distribution approximation of p value for rank product rp=9720.
righttailgamma <- function(r, k, n) 1 - pgamma(-log(r/(n+1)^k), k, scale=1)

righttailgamma(9720, 5, 500)
#--- Example 2.
# Calculate for n=500 the number of (k=)5-tuples with rankproduct rp<=9720.
# Display the exact p-value.
total <- 0
for( n in 1:9720 ) {
    if( n \%\% 1000 == 0 )
        cat("n=", n, "\\n")
    total <- total + piltzcount(n, 5, 500)
}
print(total)
total / 500^5
}

