% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{plotExprsFreqVsMean}
\alias{plotExprsFreqVsMean}
\title{Plot frequency of expression against mean expression level}
\usage{
plotExprsFreqVsMean(object, feature_set = NULL, shape = 1, alpha = 0.7,
  ...)
}
\arguments{
\item{object}{an \code{SCESet} object.}

\item{feature_set}{character, numeric or logical vector indicating a set of
features to use for the PCA. If character, entries must all be in 
\code{featureNames(object)}. If numeric, values are taken to be indices for 
features. If logical, vector is used to index features and should have length
equal to \code{nrow(object)}. If \code{NULL}, then the function checks if 
feature controls are defined. If so, then only feature controls are plotted,
if not, then all features are plotted.}

\item{shape}{(optional) numeric scalar to define the plotting shape.}

\item{alpha}{(optional) numeric scalar (in the interval 0 to 1) to define the
alpha level (transparency) of plotted points.}

\item{...}{further arguments passed to \code{\link{plotMetadata}} (should 
only be \code{size}, if anythin).}
}
\description{
Plot frequency of expression against mean expression level
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
pd <- new("AnnotatedDataFrame", data=sc_example_cell_info)
rownames(pd) <- pd$Cell
ex_sceset <- newSCESet(countData=sc_example_counts, phenoData=pd)
ex_sceset <- calculateQCMetrics(ex_sceset)
plotExprsFreqVsMean(ex_sceset)

}

