% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qc.R
\name{plotExprsMeanVsFreq}
\alias{plotExprsMeanVsFreq}
\title{Plot mean expression against frequency of expression}
\usage{
plotExprsMeanVsFreq(object, feature_set = NULL)
}
\arguments{
\item{object}{an \code{SCESet} object.}

\item{feature_set}{character, numeric or logical vector indicating a set of
features to use for the PCA. If character, entries must all be in
\code{featureNames(object)}. If numeric, values are taken to be indices for
features. If logical, vector is used to index features and should have length
equal to \code{nrow(object)}. If \code{NULL}, then the function checks if
feature controls are defined. If so, then only feature controls are plotted,
if not, then all features are plotted.}
}
\description{
Plot mean expression against frequency of expression
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
pd <- new("AnnotatedDataFrame", data=sc_example_cell_info)
rownames(pd) <- pd$Cell
ex_sceset <- newSCESet(countData=sc_example_counts, phenoData=pd)
ex_sceset <- calculateQCMetrics(ex_sceset)
plotExprsMeanVsFreq(ex_sceset)
}

