% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotMetadata}
\alias{plotMetadata}
\title{Plot metadata for cells or genes}
\usage{
plotMetadata(object, aesth = aes(x = log10(depth), y = coverage))
}
\arguments{
\item{object}{a data.frame (or object that can be coerced to such) object
containing metadata in columns to plot.}

\item{aesth}{aesthetics function call to pass to ggplot. This function
expects at least x and y variables to be supplied. The default is to plot
coverage against log10(depth).}
}
\description{
Plot metadata for cells or genes
}
\details{
Plot cell or feature metadata from an SCESet object. If one variable
 is supplied then a density plot will be returned. If both variables are
continuous (numeric) then a scatter plot will be returned. If one variable is
discrete and one continuous then a violin plot with jittered points overlaid
will be returned. If both variables are discrete then a jitter plot will be
produced. The object returned is a ggplot object, so further layers and
plotting options (titles, facets, themes etc) can be added.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
pd <- new("AnnotatedDataFrame", data = sc_example_cell_info)
example_sceset <- newSCESet(countData = sc_example_counts, phenoData = pd)
example_sceset <- calculateQCMetrics(example_sceset)
plotMetadata(pData(example_sceset))
}

