% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qc.R
\name{plotMostExpressed}
\alias{plotMostExpressed}
\title{Plot the most highly expressed genes}
\usage{
plotMostExpressed(object, col_by_variable = "coverage", n = 50,
  drop_genes = NULL)
}
\arguments{
\item{object}{an SCESet object containing expression values and
experimental information. Must have been appropriately prepared.}

\item{col_by_variable}{variable name (must be a column name of pData(object))
to be used to assign colours to cell-level values.}

\item{n}{numeric scalar giving the number of the most expressed features to
show. Default value is 50.}

\item{drop_genes}{a character, logical or numeric vector indicating which
genes (or features) to drop when producing the plot. For example, control
genes might be dropped to focus attention on contribution from biological
rather than synthetic genes.}
}
\description{
Plot the most highly expressed genes
}
\details{
Plot the percentage of reads accounted for by the top n most highly
expressed genes across the dataset.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
pd <- new("AnnotatedDataFrame", data = sc_example_cell_info)
example_sceset <- newSCESet(countData = sc_example_counts, phenoData = pd)
example_sceset <- calculateQCMetrics(example_sceset)
plotMostExpressed(example_sceset, col_by_variable="coverage")
plotMostExpressed(example_sceset, col_by_variable="Mutation_Status")
}

