% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kallisto-wrapper.R
\name{readKallistoResults}
\alias{readKallistoResults}
\title{Read kallisto results from a batch of jobs}
\usage{
readKallistoResults(kallisto_log = NULL, samples = NULL,
  directories = NULL, read_h5 = FALSE, verbose = TRUE)
}
\arguments{
\item{kallisto_log}{list, generated by \code{runKallisto}. If provided, then
\code{samples} and \code{directories} arguments are ignored.}

\item{samples}{character vector providing a set of sample names to use for
the abundance results.}

\item{directories}{character vector providing a set of directories containing
kallisto abundance results to be read in.}

\item{read_h5}{logical, should the bootstrap results be read in from the HDF5
objects produced by kallisto?}

\item{verbose}{logical, should function provide output about progress?}
}
\description{
After generating transcript/feature abundance results using kallisto for a
batch of samples, read these abundance values into an \code{SCESet} object.
}
\details{
This function expects to find only one set of kallisto abundance
results per directory; multiple adundance results in a given directory will
be problematic.
}
\examples{
\dontrun{
kallisto_log <- runKallisto("targets.txt", "transcripts.idx", single_end=FALSE,
         output_prefix="output", verbose=TRUE, n_bootstrap_samples=10)
sceset <- readKallistoResults(kallisto_log)
}
}

