% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kallisto-wrapper.R
\name{readKallistoResultsOneSample}
\alias{readKallistoResultsOneSample}
\title{Read kallisto results for a single sample into a list}
\usage{
readKallistoResultsOneSample(directory, read_h5 = FALSE)
}
\arguments{
\item{directory}{character string giving the path to the directory containing
the kallisto results for the sample.}

\item{read_h5}{logical, if \code{TRUE} then read in bootstrap results from
the HDF5 object produced by kallisto.}
}
\value{
A list with two elements: (1) a data.frame \code{abundance} with
columns for 'target_id' (feature, transcript, gene etc), 'length' (feature
length), 'eff_length' (effective feature length), 'est_counts' (estimated
feature counts), 'tpm' (transcripts per million) and possibly many columns
containing bootstrap estimated counts; and (2) a list \code{run_info} with
details about the kallisto run that generated the results.
}
\description{
Read kallisto results for a single sample into a list
}
\details{
The directory is expected to contain results for just a single
sample. Putting more than one sample's results in the directory will result
in unpredictable behaviour with this function. The function looks for the
files (with the default names given by kallisto) 'abundance.txt',
'run_info.json' and (if \code{read_h5=TRUE}) 'abundance/h5'. If these files
are missing, or if results files have different names, then this function
will not find them.
}
\examples{
# If kallisto results are in the directory "output", then call:
# readKallistoResultsOneSample("output")
}

