% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scaleAmp.R
\name{scaleAmp}
\alias{scaleAmp}
\title{Scale expression values to have average amplitude of 1 in
expressing cells}
\usage{
scaleAmp(expr_matrix, is_expr)
}
\arguments{
\item{expr_matrix}{a numeric matrix containing expression values
counts.}

\item{is_expr}{a logical matrix of same dimensions as
\code{expr_matrix} indicating which cells are expressed for each
gene.}
}
\value{
A matrix object.
}
\description{
Scale expression values to have average amplitude of 1 in
expressing cells
}
\details{
This function is intended to process RNA-Seq or ChIP-Seq
data prior to modeling with RankProduct. Written by Davis McCarthy.
}

