\name{scater-package}
\alias{scater-package}
\alias{scater}
\docType{package}
\title{
  Single-cell analysis toolkit for gene expression in R
}
\description{
  A collection of tools for differential analysis of single-cell gene
  expression data. Leaning heavily on packages like limma, this package
  implements some simple approaches to studying expression frequency,
  expression amplitude, expression variability and coexpression in
  RNA-seq data from single cells.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Davis McCarthy <davismcc@gmail.com>.

Maintainer: Davis McCarthy <davismcc@gmail.com>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
% Optionally other standard keywords, one per line,
% from the file KEYWORDS in the R documentation.
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  # Optional simple examples of the most important functions
}
