% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scater.R
\docType{package}
\name{scater}
\alias{scater}
\alias{scater-package}
\title{Single-cell analysis toolkit for expression in R}
\description{
\pkg{scater} provides a class and numerous functions for the analysis and
visualisation of single-cell RNA-seq expression data.
}
\details{
In particular, \pkg{scater} provides rank-based methods for finding
differentially expressed genes. Testing for DE genes using the
rank-product method can be done with the function \code{\link{testRankProd}}.
}

