% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.R
\name{scater_gui}
\alias{scater_gui}
\title{scater GUI function}
\usage{
scater_gui(sce_set)
}
\arguments{
\item{sce_set}{SCESet object after the calculateQCMetrics()}
}
\value{
Opens a browser window with an interactive shiny app and visualize
all possible plots included in the scater
}
\description{
scater shiny app GUI for workflow for less programmatically inclined users
}

