% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summariseExprsAcrossFeatures.R
\name{summariseExprsAcrossFeatures}
\alias{summariseExprsAcrossFeatures}
\title{Summarise expression values across feature}
\usage{
summariseExprsAcrossFeatures(object, use_as_exprs = "tpm",
  summarise_by = "feature_id")
}
\arguments{
\item{object}{an \code{SCESet} object.}

\item{use_as_exprs}{character string indicating which slot of the
assayData from the \code{SCESet} object should be used as expression values.
Valid options are \code{'exprs'} the expression slot, \code{'tpm'} the
transcripts-per-million slot or \code{'fpkm'} the FPKM slot.}

\item{summarise_by}{character string giving the column of \code{fDat(object)}
that will be used as the features for which summarised expression levels are
to be produced. Default is \code{'feature_id'}.}
}
\description{
Create a new \code{SCESet} with counts summarised at a different feature
level. A typical use would be to summarise transcript-level counts at gene
level.
}
\details{
Only transcripts-per-million (TPM) and fragments per kilobase of
exon per million reads mapped (FPKM) expression values should be aggregated
across features. Since counts are not scaled by the length of the feature,
expression in counts units are not comparable within a sample without
adjusting for feature length. Thus, we cannot sum counts over a set of
features to get the expression of that set (for example, we cannot sum counts
 over transcripts to get the expression for a gene). See the following link
 for a discussion of RNA-seq expression units by Harold Pimentel:
\url{https://haroldpimentel.wordpress.com/2014/05/08/what-the-fpkm-a-review-rna-seq-expression-units/}.
}
\examples{
\dontrun{
object <- summariseExprsAcrossFeatures(object)
}
}

