% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/testAmp.R
\name{testAmp}
\alias{testAmp}
\title{Test for differences in expression amplitude}
\usage{
testAmp(object, design, coef = ncol(design), expr_values = "cpm",
  method = "ttest", p_adj_method = "BH", ...)
}
\description{
Test for differences in expression amplitude
}
\details{
A user-level wrapper function for different approaches to
testing for differences in expression amplitude. Methods
implemented are: t-test, rank-production test, ...
}

