% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/testAmp.R
\name{testAmp.ttest.onegene}
\alias{testAmp.ttest.onegene}
\title{Test for differences in expression for one gene with a t-test}
\usage{
testAmp.ttest.onegene(log2cpm, design, test_variable = ncol(design))
}
\arguments{
\item{log2cpm}{vector of log2-count-per-million values, the measure of
expression}

\item{design}{design matrix for the linear model fit}

\item{test_variable}{the variable (or coefficient) for which testing is done}
}
\value{
dataframe with logFC, SE, N (number of cells expressed), t-statistic
and p-value
}
\description{
Test for differences in expression for one gene with a t-test
}
\examples{
x <- rnorm(10)
design <- model.matrix(~factor(rep(c(1,2), each = 5)) + factor(rep(c(1,2), 5)))
testAmp.ttest.onegene(x, design)
}
\seealso{
\code{\link{nchar}} which this function wraps
}

