% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/testAmp.R
\name{testAmp.ttest}
\alias{testAmp.ttest}
\title{Test for differences in expression amplitude using a t-test}
\usage{
testAmp.ttest(object, design, coef = ncol(design),
  expr_values = "log2cpm.pc0.25", p_adj_method = "BH")
}
\arguments{
\item{object}{an SCESet object containing expression values and
experimental information. Must have been appropriately prepared.}

\item{design}{design matrix describing experimental design}

\item{coef}{coefficient(s) of the design matrix for which testing is to be
done}

\item{expr_values}{character string giving the name of the element
of the SCESet object yielding the expression values to be used.}

\item{p_adj_method}{method to be used to adjust p-values for multiple testing
(defaults to Benjamini-Hochberg FDR)}
}
\description{
Test for differences in expression amplitude using a t-test
}

