% Generated by roxygen2 (4.0.2): do not edit by hand
\name{testRankProduct.matrix}
\alias{testRankProduct.matrix}
\title{Conduct RankProduct test for a data matrix}
\usage{
testRankProduct.matrix(data_matrix, group, block = NULL,
  delta = "geometric")
}
\arguments{
\item{data_matrix}{numeric matrix with expression values
(log-scale) for features (rows) and samples/cells (columns).}

\item{group}{character vector or factor defining the two groups to
be compared}

\item{block}{character vector or factor defining the experimental
blocks within which the two groups are to be compared}

\item{delta}{character string indication approximate p-value to be
returned for rank-product results: "upper", "lower", or "geometric"
(default; gives approximate p-value)}
}
\value{
results object summarising RankProduct results
}
\description{
Conduct RankProduct test for a data matrix
}

