% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/testRankProduct.R
\name{testRankProduct}
\alias{testRankProduct}
\title{Conduct rank-product test}
\usage{
testRankProduct(dfilt, nmax.genes = 500, max.rprod.exact = 1e+05)
}
\arguments{
\item{dfilt}{DGEList object containing data and so on that has been
filtered and expressed cells determined}

\item{nmax.genes}{integer giving the maximum number of genes for
which to compute exact rank-product p-values}

\item{max.rprod.exact}{integer giving the maximum rank-product for
which to compute exact p-values. Exact p-values are very slow to
compute for large rank-products, and converge to approximate
p-values.}
}
\value{
data frame with mean rank, rank product, number of ways of
obtaining given rank product, approximate p-value and exact p-value
}
\description{
Conduct rank-product test
}

