% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rank_product_exact.R
\name{tidydivisor}
\alias{tidydivisor}
\title{Obtain all divisors of a set of prime numbers and their powers}
\usage{
tidydivisor(f.and.m)
}
\arguments{
\item{f.and.m}{matrix in which the first column gives the primes and the second column gives the prime powers, e.g. the output of \code{tidyfactor}}
}
\value{
sorted vector of divisors of $f_1^m_1 \times f_2^m_2 \times f_3^m_3 \ldots$
}
\description{
Obatin all divisors of $f_1^m_1 \times f_2^m_2 \times f_3^m_3 \ldots$, where $f_1, f_2, f_3 \ldots$ are primes and $m_1, m_2, m_3 \ldots$ are prime powers.
}
\examples{
f.and.m <- tidyfactor(281137500)
tidydivisor(f.and.m)
}
\seealso{
tidyfactor
}
\keyword{arithmetic}

