% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rank_product_exact.R
\name{tidyfactor}
\alias{tidyfactor}
\title{Extract prime factors and powers of rank product.}
\usage{
tidyfactor(n)
}
\arguments{
\item{n}{integer to be factorized}
}
\value{
data frame in which each row gives a unique prime factor and its power
}
\description{
Given an integer (in this context a rank product), factorize it into its prime factors and their powers.
}
\examples{
tidyfactor(281137500)
}
\keyword{arithmetic}

