% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SCESet-methods.R
\name{toCellDataSet}
\alias{toCellDataSet}
\title{Convert an \code{SCESet} to a \code{CellDataSet}}
\usage{
toCellDataSet(sce, use_as_exprs = "exprs")
}
\arguments{
\item{sce}{An \code{SCESet} object}

\item{use_as_expr}{What should \code{exprs(cds)} be mapped from in the \code{SCESet}? Should be
one of "exprs", "tpm", "fpkm", "counts"}
}
\value{
An object of class \code{SCESet}
}
\description{
Convert an \code{SCESet} to a \code{CellDataSet}
}

