% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helpers.R
\name{trim}
\alias{trim}
\title{#' Trim whitespace from start and end of a (vector of) string(s)}
\usage{
trim(x)
}
\arguments{
\item{x}{vector of strings (of length one or greater)}
}
\value{
vector of strings with whitespace trimmed from start and end of each string
}
\description{
#' Trim whitespace from start and end of a (vector of) string(s)
}
\examples{
trim("   This is a string   ")
}

