% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rank_product_exact.R
\name{uniquerows}
\alias{uniquerows}
\title{Get the unique rows of a matrix}
\usage{
uniquerows(A)
}
\arguments{
\item{A}{matrix for which we obtain unique rows}
}
\value{
A list containing the matrix C, containing only the unique rows of A, and an index vector \code{ia} giving the indices of the unique rows of A in C.
}
\description{
Save the unique rows of input matrix A in a new matrix C and create
an index vector ia, containing the sequence numbers of the rows in A
that were saved in C.
}
\examples{
A <- matrix(rep(1:4, each = 4), byrow = TRUE, ncol = 2)
uniquerows(A)
}
\keyword{manip}

