% Generated by roxygen2 (4.0.2): do not edit by hand
\name{updateParam}
\alias{updateParam}
\title{Update parameters for computation of rank-product p-values}
\usage{
updateParam(param, n, k, j, delta)
}
\arguments{
\item{param}{a matrix of lists, where each element of param is a
list with values for the parameters a through e; these parameters
specify the functional form of the bound; a, b, and c are all
vectors of unknown length, d and e are scalars.}

\item{n}{the number of features (e.g. genes).}

\item{k}{the number of replicates for which we need to compute the
corresponding parameters.}

\item{j}{the index of the interval for which we need to compute the
corresponding parameters.}

\item{delta}{0 for the lower bound and 1 for the upper bound.}
}
\value{
A possibly updated set of parameters, at least including
those corresponding to (k,j).
}
\description{
This subroutine updates the current set of parameters to make sure
that the parameters corresponding to k replicates and the j'th
interval are included.
}
\details{
This subroutine makes sure that the parameters
corresponding to k replicates and a rank product within the j'th
interval are already present. If they already are (because
calculated before), it does not compute anything. Otherwise, it
recursively computes all parameters that are needed to arrive at
the parameters for (k,j).

This implementation closely follows HEB, in particular equations (9) through (11).
}

