% Generated by roxygen2 (4.0.2): do not edit by hand
\name{voomsc}
\alias{voomsc}
\title{A version of voom() from limma tweaked for single-cell expression data}
\usage{
voomsc(counts, design = NULL, lib.size = NULL, normalize.method = "none",
  plot = FALSE, span = 0.5, prior.count = 0.5, ...)
}
\arguments{
\item{counts}{a numeric matrix containing raw counts, or an
ExpressionSet containing raw counts, or a DGEList object.}

\item{design}{design matrix with rows corresponding to samples and
columns to coefficients to be estimated. Defaults to the unit
vector meaning that samples are treated as replicates.}

\item{lib.size}{numeric vector containing total library sizes for
each sample. If \code{NULL} and counts is a DGEList then, the
normalized library sizes are taken from counts.  Otherwise library
sizes are calculated from the columnwise counts totals.}

\item{normalize.method}{numeric vector containing total library
sizes for each sample. If NULL and counts is a DGEList then, the
normalized library sizes are taken from counts.  Otherwise library
sizes are calculated from the columnwise counts totals.}

\item{plot}{logical, should a plot of the mean-variance trend be displayed?}

\item{span}{width of the lowess smoothing window as a proportion.}

\item{prior.count}{offset, on the scale of raw counts, to be added
to raw counts before converting to log2(counts-per-million);
default is 0.5.}
}
\value{
An EList object.
}
\description{
A version of voom() from limma tweaked for single-cell expression data
}
\details{
This function is intended to process RNA-Seq or ChIP-Seq
data prior to linear modelling in limma, written by Charity Law and
Gordon Smyth. voom is an acronym for mean-variance modelling at the
observational level. The key concern is to estimate the
mean-variance relationship in the data, then use this to compute
appropriate weights for each observation. Count data almost show
non-trivial mean-variance relationships. Raw counts show increasing
variance with increasing count size, while log-counts typically
show a decreasing mean-variance trend. This function estimates the
mean-variance trend for log-counts, then assigns a weight to each
observation based on its predicted variance. The weights are then
used in the linear modelling process to adjust for
heteroscedasticity.  In an experiment, a count value is observed
for each tag in each sample. A tag-wise mean-variance trend is
computed using lowess. The tag-wise mean is the mean log2 count
with a (default) offset of 0.5, across samples for a given tag. The
tag-wise variance is the quarter-root-variance of normalized2 log2
counts per million values with a (default) offset of 0.5, across
samples for a given tag. Tags with zero counts across all samples
are not included in the lowess fit. Optional normalization is
performed using normalizeBetweenArrays. Using fitted values of log2
counts from a linear model fit by lmFit, variances from the
mean-variance trend were interpolated for each observation. This
was carried out by approxfun. Inverse variance weights can be used
to correct for mean-variance trend in the count data.
}

