% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SingleCellExperiment-methods.R
\docType{methods}
\name{norm_exprs}
\alias{norm_exprs}
\alias{norm_exprs<-}
\alias{stand_exprs}
\alias{stand_exprs<-}
\alias{fpkm}
\alias{fpkm<-}
\alias{accessors exprs norm_exprs stand_exprs fpkm}
\alias{exprs}
\alias{exprs,SingleCellExperiment-method,}
\alias{exprs<-,SingleCellExperiment,matrix-method}
\title{Additional accessors for the typical elements of a SingleCellExperiment object.}
\usage{
norm_exprs(object)

norm_exprs(object) <- value

stand_exprs(object)

stand_exprs(object) <- value

fpkm(object)

fpkm(object) <- value

\S4method{exprs}{SingleCellExperiment}(object)
}
\arguments{
\item{object}{SingleCellExperiment class object from which to access or to 
which to assign assay values. Namely: "exprs", norm_exprs", "stand_exprs", "fpkm".
The following are imported from \code{\link{SingleCellExperiment}}: "counts",
"normcounts", "logcounts", "cpm", "tpm".}

\item{value}{a numeric matrix (e.g. for \code{exprs})}

\item{object}{a \code{SingleCellExperiment} object.}
}
\value{
a matrix of normalised expression data

a matrix of standardised expressiond data

a matrix of FPKM values

A matrix of numeric, integer or logical values.
}
\description{
Convenience functions to access commonly-used assays of the 
\code{\link{SingleCellExperiment}} object.
}
\details{
\code{exprs} values are typically log2-scaled (normalised) 
counts-per-million values.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
example_sce <- SingleCellExperiment(
assays = list(counts = sc_example_counts), colData = sc_example_cell_info)

exprs(example_sce) <- log2(calculateCPM(example_sce, use.size.factors = FALSE) + 1)

example_sce <- SingleCellExperiment(
assays = list(norm_counts = sc_example_counts), colData = sc_example_cell_info)

counts(example_sce) <- sc_example_counts
norm_exprs(example_sce) <- log2(calculateCPM(example_sce, use.size.factors = FALSE) + 1)

stand_exprs(example_sce) <- log2(calculateCPM(example_sce, use.size.factors = FALSE) + 1)

tpm(example_sce) <- calculateTPM(example_sce, effective_length = 5e4)

cpm(example_sce) <- calculateCPM(example_sce, use.size.factors = FALSE)

fpkm(example_sce)
}
\author{
Davis McCarthy
}
