% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/testFreq.R
\name{binomGLMFit}
\alias{binomGLMFit}
\title{Test for differences in expression frequency using logistic regression}
\usage{
binomGLMFit(y, design, ntest = nrow(y))
}
\arguments{
\item{y}{a matrix containing binary data on whether a
gene is expressed or not in each cell, with rows corresponding to
genes and columns to cells.}

\item{design}{a design matrix to be used for the GLM fit.}

\item{ntest}{the number of tests to carry out (default: test all genes).}
}
\description{
Test for differences in expression frequency using logistic regression
}
\details{
Fit a binomial GLM for expression frequency to each gene.
}

