% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\name{bplot}
\alias{bplot}
\title{Beeswarm plot for expression levels}
\usage{
bplot(expr_values, group = NULL, ylab = expression("log2(cpm)"),
  de_direction = NULL, distribution_plot = "violinplot",
  col_theme = "few")
}
\arguments{
\item{expr_values}{matrix of expression values to be plotted.}

\item{group}{vector or factor that defines groups to be coloured
disctinctly in the plot.}

\item{ylab}{label for the y-axis.}

\item{de_direction}{optional character vector indicating whether genes are up-
or down-regulated (or giving any other information to be plotted
above beeswarms).}

\item{distribution_plot}{show either a violinplot or a boxplot
behind the beeswarm? (Default = \code{"violinplot"}).}
}
\description{
\code{bplot} creates a beeswarm plot with background boxplots to
show the distribution of expression values for a set of genes.
}

