% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calcRankProdPvalsApprox}
\alias{calcRankProdPvalsApprox}
\title{Right-tailed gamma approximation to rank-product p-value}
\usage{
calcRankProdPvalsApprox(r, k, n)
}
\arguments{
\item{r}{integer rank-product}

\item{k}{integer number of experiments}

\item{n}{integer number of genes}
}
\value{
float aproximate p-value
}
\description{
Right-tailed gamma approximation to rank-product p-value
}
\examples{
calcRankProdPvalsApprox(9720, 5, 500)
}

