% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qc.R
\name{calculateQCMetrics}
\alias{calculateQCMetrics}
\title{Calculate QC metrics}
\usage{
calculateQCMetrics(object, feature_controls = NULL, cell_controls = NULL,
  nmads = 5, pct_feature_controls_threshold = 80)
}
\arguments{
\item{object}{an SCESet object containing expression values and
experimental information. Must have been appropriately prepared.}

\item{feature_controls}{a character vector of feature names, or a logical vector,
or a numeric vector of indices used to identify feature controls (for example,
ERCC spike-in genes, mitochondrial genes, etc).}

\item{cell_controls}{a character vector of cell (sample) names, or a logical
vector, or a numeric vector of indices used to identify cell controls (for
example, blank wells or bulk controls).}

\item{nmads}{numeric scalar giving the number of median absolute deviations to be
used to flag potentially problematic cells based on depth (total number of
counts for the cell, or library size) and coverage (number of features with
non-zero expression). Default value is 5.}

\item{pct_feature_controls_threshold}{numeric scalar giving a threshold for
percentage of expression values accounted for by feature controls. Used as to
flag cells that may be filtered based on high percentage of expression from
feature controls.}
}
\description{
Calculate QC metrics
}
\details{
Calculate useful quality control metrics to help with pre-processing
of data and identification of potentially problematic features and cells.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
pd <- new("AnnotatedDataFrame", data=sc_example_cell_info)
rownames(pd) <- pd$Cell
example_sceset <- newSCESet(countData=sc_example_counts, phenoData=pd)
example_sceset <- calculateQCMetrics(example_sceset)
}

