% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SCESet-methods.R
\name{fromCellDataSet}
\alias{fromCellDataSet}
\title{Convert a \code{CellDataSet} to an \code{SCESet}}
\usage{
fromCellDataSet(cds, useExpression = TRUE, logged = FALSE)
}
\arguments{
\item{cds}{A \code{CellDataSet} from the \code{monocle} package}

\item{useExpression}{logical If TRUE (default), `cellData` is mapped to `exprs(sce)`, otherwise `counts(sce)`}

\item{logged}{logical, if a value is supplied for the cellData argument, are the expression values already on the log2 scale, or not?}
}
\value{
An object of class \code{SCESet}
}
\description{
Convert a \code{CellDataSet} to an \code{SCESet}
}

