% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getLogFCOneBlock.R
\name{getLogFCOneBlock}
\alias{getLogFCOneBlock}
\title{Get logFC values for replicates within one experimental block}
\usage{
getLogFCOneBlock(data_matrix, group)
}
\arguments{
\item{data_matrix}{numeric matrix with expression values
(log-scale) for features (rows) and samples/cells (columns).}

\item{group}{character vector or factor defining the two groups to
be compared}
}
\value{
matrix of logFC values for all pairwise comparisons of
samples in group1 and group2
}
\description{
Get logFC values for replicates within one experimental block
}

