% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getRanksOneBlock.R
\name{getRanksOneBlock}
\alias{getRanksOneBlock}
\title{Get ranks from logFCs for one block}
\usage{
getRanksOneBlock(logFC_matrix, group, logged = FALSE)
}
\arguments{
\item{logFC_matrix}{numeric matrix of logFCs computed as group1 -
group2}

\item{group}{character vector or factor defining the experimental
groups being compared}

\item{logged}{boolean: return raw ranks if \code{FALSE} or
log(ranks) if \code{TRUE}}
}
\description{
Get ranks both for up-regulation in group1 and up-regulation in
group2
}

