% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qc.R
\name{plotExplanatoryVariables}
\alias{plotExplanatoryVariables}
\title{Plot explanatory variables ordered by percentage of phenotypic variance explained}
\usage{
plotExplanatoryVariables(object, variables = c("coverage", "depth"))
}
\arguments{
\item{object}{an SCESet object containing expression values and
experimental information. Must have been appropriately prepared.}

\item{variables}{vectory of variable names (must be columns of pData(object))
to be plotted.}
}
\description{
Plot explanatory variables ordered by percentage of phenotypic variance explained
}
\details{
Function produces a pairs plot of the explanatory variables ordered
by the percentage of gene expression variance (as measured by R-squared in a
marginal linear model) explained by variable. Median percentage R-squared is
reported on the plot for each variable. Discrete variables are coerced to a
factor and plotted as integers with jittering. Variables with only one unique
value are quietly ignored.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
pd <- new("AnnotatedDataFrame", data = sc_example_cell_info)
rownames(pd) <- pd$Cell
example_sceset <- newSCESet(countData = sc_example_counts, phenoData = pd)
example_sceset <- calculateQCMetrics(example_sceset)
vars <- names(pData(example_sceset))[c(2:3, 5:14)]
plotExplanatoryVariables(example_sceset, variables=vars)
}

