% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotExpression}
\alias{plotExpression}
\title{Plot expression values for a set of genes or features}
\usage{
plotExpression(features, data_object, aesth, ncol = 2, xlab = "",
  ylab = "log2(counts-per-million + 1)", show_median = FALSE,
  show_violin = FALSE)
}
\arguments{
\item{features}{a character vector of feature names or Boolean
vector indicating which features should have their expression
values plotted}

\item{data_object}{an SCESet object containing expression values and
experimental information. Must have been appropriately prepared.}

\item{aesth}{aesthetics function call to pass to ggplot}

\item{ncol}{number of columns to be used for the panels of the plot}

\item{xlab}{label for x-axis}

\item{ylab}{label for y-axis}

\item{show_median}{logical, show the median for each group on the plot}

\item{show_violin}{logical, show a violin plot for the distribution
for each group on the plot}
}
\description{
Plot expression values for a set of genes or features
}
\details{
Plot expression values (default log2(counts-per-million +
1)) for a set of genes or features.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
pd <- new("AnnotatedDataFrame", data = sc_example_cell_info)
example_sceset <- newSCESet(countData = sc_example_counts, phenoData = pd)
example_sceset <- calculateQCMetrics(example_sceset)
plotExpression(1:5, example_sceset,
aes(x=Mutation_Status, y=log2_evals, colour=Patient), xlab="Mutation_Status",
show_violin=TRUE, show_median=TRUE)
}

