% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotExpression}
\alias{plotExpression}
\title{Plot expression values for a set of genes or features}
\usage{
plotExpression(features, data_object, aes, ncol = 2, xlab = "Patient",
  ylab = "log2(counts-per-million + 1)", show_median = FALSE,
  show_violin = FALSE)
}
\arguments{
\item{features}{a character vector of feature names or Boolean
vector indicating which features should have their expression
values plotted}

\item{data_object}{a DGEList object containing expression values and
experimental information. Must have been appropriately prepared.}

\item{aes}{aesthetics function call to pass to ggplot}

\item{ncol}{number of columns to be used for the panels of the plot}

\item{show_median}{logical, show the median for each group on the plot}

\item{show_violin}{logical, show a violin plot for the distribution
for each group on the plot}
}
\description{
Plot expression values for a set of genes or features
}
\details{
Plot expression values (default log2(counts-per-million +
1)) for a set of genes or features.
}

