% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/testAmp.R
\name{testAmp}
\alias{testAmp}
\title{Test for differences in expression amplitude}
\usage{
testAmp(object, design, coef = ncol(design), expr_values = "cpm",
  method = "ttest", p_adj_method = "BH", ...)
}
\arguments{
\item{object}{SCESet class object containing data}

\item{design}{design matrix describing experimental design}

\item{coef}{coefficient(s) of the design matrix for which testing is to be
done}

\item{expr_values}{which expression values to use?}

\item{method}{method to be used to test for differences in expression
amplitude (default is a t-test)}

\item{p_adj_method}{method to be used to adjust p-values for multiple testing
(defaults to Benjamini-Hochberg FDR)}

\item{...}{arguments passed to \code{testAmp.ttest}}
}
\description{
Test for differences in expression amplitude
}
\details{
A user-level wrapper function for different approaches to
testing for differences in expression amplitude. Methods
implemented are: t-test, rank-production test, ...
}

