% Generated by roxygen2 (4.0.1): do not edit by hand
\name{testAmp.ttest}
\alias{testAmp.ttest}
\title{Test for differences in expression amplitude using a t-test}
\usage{
testAmp.ttest(object, expr.values = "log2cpm.pc0.25",
  experiment = "culture", treatment = "perturbed", p.adj.method = "BH")
}
\arguments{
\item{object}{a DGEList object containing expression values and
experimental information. Must have been appropriately prepared.}

\item{expr.values}{character string giving the name of the element
of the DGEList object yielding the expression values to be used.}
}
\description{
Test for differences in expression amplitude using a t-test
}

