% Generated by roxygen2 (4.0.2): do not edit by hand
\name{testAmp.ttest}
\alias{testAmp.ttest}
\title{Test for differences in expression amplitude using a t-test}
\usage{
testAmp.ttest(object, design, coef = ncol(design),
  expr_values = "log2cpm.pc0.25", p.adj.method = "BH")
}
\arguments{
\item{object}{a DGEList object containing expression values and
experimental information. Must have been appropriately prepared.}

\item{expr_values}{character string giving the name of the element
of the DGEList object yielding the expression values to be used.}
}
\description{
Test for differences in expression amplitude using a t-test
}

