% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/testFreq.R
\name{testFreq}
\alias{testFreq}
\title{Test for differences in expression frequency using logistic regression}
\usage{
testFreq(data_object, design, coef = ncol(design), p_adj_method = "BH",
  ntest = nrow(data_object))
}
\arguments{
\item{data_object}{a DGEList object containing expression values and
experimental information. Must have been appropriately prepared.}

\item{design}{a design matrix to be used for the GLM fit.}

\item{coef}{the coefficient(s)}

\item{p_adj_method}{method to pass to \code{p.adjust()} to adjust
p-values for multiple testing.}

\item{ntest}{number of tests to carry out (defaults to number of genes)}
}
\description{
Test for differences in expression frequency using logistic regression
}
\details{
Test for differences in expression frequency using
logistic regression and a likelihood ratio test. Returns a DGELRT
object with the differential frequency test results.
}

