% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/testRankProduct.R
\name{testRankProduct.matrix}
\alias{testRankProduct.matrix}
\title{Conduct RankProduct test for a data matrix}
\usage{
testRankProduct.matrix(data_matrix, group, block = NULL,
  delta = "geometric", compute_pvals = FALSE)
}
\arguments{
\item{data_matrix}{numeric matrix with expression values
(log-scale) for features (rows) and samples/cells (columns).}

\item{group}{character vector or factor defining the two groups to
be compared}

\item{block}{character vector or factor defining the experimental
blocks within which the two groups are to be compared}

\item{delta}{character string indication approximate p-value to be
returned for rank-product results: "upper", "lower", or "geometric"
(default; gives approximate p-value)}

\item{compute_pvals}{logical, should p-values be computed? Default
is \code{FALSE} as for experiments with many genes or cells,
p-values will be extremely expensice or impossible to compute}
}
\value{
results object summarising RankProduct results
}
\description{
Conduct RankProduct test for a data matrix
}

