% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SCESet-methods.R
\name{toCellDataSet}
\alias{toCellDataSet}
\title{Convert an \code{SCESet} to a \code{CellDataSet}}
\usage{
toCellDataSet(sce, useExpression = TRUE)
}
\arguments{
\item{sce}{An \code{SCESet} object}

\item{useExpression}{If TRUE (default), `exprs(sce)` is used as the `exprsData`, otherwise `counts(sce)`}
}
\value{
An object of class \code{SCESet}
}
\description{
Convert an \code{SCESet} to a \code{CellDataSet}
}

